--- src/features/application-logs/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as ApplicationLogsForm } from './ApplicationLogsForm';
