+++ src/features/send-application-logs/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as SendApplicationLogsForm } from './SendApplicationLogsForm';
